Fix services listing

+++ lib/pages/market/selected_listing_page.dart
@@ -116,7 +116,7 @@ class _SelectedListingPageState extends ConsumerState<SelectedListingPage> {
                                 style:
                                     Theme.of(context).textTheme.headlineSmall!),
                             RatingBarIndicator(
-                              rating: listing.rating!,
+                              rating: listing.rating!.toDouble(),
                               itemBuilder: (context, index) {
                                 return Icon(
                                   Icons.star_rounded,
